refactor functions into cores
