+++ multiplex_immuno_processing/find_matched_positions_across_rounds.py
@@ -14,6 +14,8 @@ import pdb
 import xml.etree.ElementTree as ET
 import lxml.etree as etree
 import re
+from core.utils import get_round_info_from_dict, get_available_positions, find_matching_position_scene
+
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--input_yaml", type=str, required=True, help="yaml config path")
@@ -21,76 +23,6 @@ parser.add_argument("--refrence_round", type=str, default="R1", required=False, 
 
 # For each postiion
 # have a dataframe with columns: Round, scene, align_channel, Refrence_round(True/False)
-
-
-class czi_metadata_helper():
-    # TODO: modify with othere get functions that return other useful metadata... currently set up to only return position/scene number but can change that in the future
-    def __init__(self, czi_filepath):
-        self.img = AICSImage(czi_filepath)
-        metadata_raw = self.img.metadata
-        self.metastr = ET.tostring(metadata_raw).decode("utf-8")
-        self.metadata = etree.fromstring(self.metastr)
-
-    def get_well_id_from_scene(self, scene):
-        """From xml element, extract well id"""
-        shape = scene.find(".//Shape")
-        well_id = shape.get("Name")
-        # zero pad the well_id number by two
-        non_numeric_part = ''.join(filter(lambda x: not x.isdigit(), well_id))
-        numeric_part = ''.join(filter(lambda x: x.isdigit(), well_id))
-        well_id_str = non_numeric_part + f"{str(numeric_part).zfill(2)}"
-        return well_id_str
-
-    def debug_save_metadata_as_xml(self):
-        """For debuggging purposes, we save the metadata as xml"""
-        with open("output_xml_debug.xml", "w") as xml_file: 
-            xml_file.write(self.metastr)
-        
-        
-    def get_position_scene_wellid_paired_list(self):
-        scenes = self.metadata.findall(".//Scenes/Scene")
-        print(len(scenes))
-        scenes_list = []
-        positions_list = []
-        well_ids_list = []
-
-        for scene in scenes:
-            scene_index = scene.get("Index")
-            scene_name = scene.get("Name")
-            center_position = scene.find(".//CenterPosition").text
-            well_id = self.get_well_id_from_scene(scene)
-            scenes_list.append(int(scene_index)+1)
-            positions_list.append(int(re.findall(r'\d+', scene_name)[0]))
-            well_ids_list.append(well_id)
-
-        return positions_list, scenes_list, well_ids_list
-
-
-
-
-def get_round_info_from_dict(round_of_intrest, dataset):
-    """Returns the information for the round of intrest from a dict that is structured according to our yaml config file"""
-    round_info= [dataset['Data'][f] for f in range(len(dataset['Data'])) if dataset['Data'][f]['round'] == round_of_intrest]
-    assert len(round_info)==1, "Multiple rounds with the same name found or none found"
-    return round_info[0]
-
-def get_available_positions(round_info):
-    """loads the czi and ckecks the list of positions available"""
-    #dataframe = zen_position_helper.get_position_info_from_czi(round_info['path'])
-    czi_file = czi_metadata_helper(round_info['path'])
-    positions, scenes, well_ids = czi_file.get_position_scene_wellid_paired_list()
-    assert len(positions) == len(scenes) == len(well_ids)
-    return positions, scenes, well_ids
-
-def find_matching_position_scene(all_positions_in_round, all_scenes_in_round, well_ids_all, refrence_postion):
-    """returns the corresponding position, scene pair that matches refrence position in refrence round"""
-    position_index = all_positions_in_round.index(refrence_postion)
-    matching_scene = all_scenes_in_round[position_index]
-    well_id_match = well_ids_all[position_index]
-
-    return all_positions_in_round[position_index], matching_scene, well_id_match
-
-
 
 class create_registration_matching_dataset():
     def __init__(self, yaml_file, refrence_round):
